refactor: simplify inertia client bootstrap
